fix(snapshot): don't rely on semantic versioning (#12)
This pull request proposes a change in `deltavim` approach to snapshot. Instead of relying on plugins to follow semantic versioning, which is often inconsistent, it will pin known working commits. 

That is, whereas plugins should adhere to semantic versioning by incrementing major, minor, and patch versions, in practice, it’s quite rare for them to comply:

- About half of the plugins may start with the intention to follow semantic versioning but then abandon it midway.
- The other half might not make any effort to follow it at all.
- Even when plugins do attempt semantic versioning, they often fail to do so correctly and end up violating the specification.

Instead, let's opt for a more reliable approach: pinning known working commits. This has  proven to be more dependable for snapshotting.